% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{digest}
\alias{digest}
\title{Create ironseed digests for arbitrary R objects}
\usage{
digest(object, n = 1L, salt = 1L, serialize = TRUE, ascii = FALSE, xdr = FALSE)
}
\arguments{
\item{object}{an arbitary R object. Will be serialized unless the
\code{serialize} argument is \code{FALSE}.}

\item{n}{a scalar integer. Specifies the length of the returned vector.}

\item{salt}{a scalar integer. Used to vary the digests between applications.}

\item{serialize}{a logical. Indicates whether to serialize object before
calculating the digest.}

\item{ascii, xdr}{Passed to \link[base:serialize]{base::serialize}}
}
\value{
an integer vector of length \code{n}
}
\description{
The \code{digest()} function creates a variable length ironseed digest for
arbitrary R objects. Internally, it uses \link[base:serialize]{base::serialize} to convert objects
to a raw vector before calculating the digest.
}
\details{
This algorithm uses different coefficients than \link{create_ironseed} and
\link{create_seedseq}.
}
