% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ironseed.R
\name{fill_random_seed}
\alias{fill_random_seed}
\title{Initialize .Random.seed}
\usage{
fill_random_seed(x, quiet = FALSE)
}
\arguments{
\item{x}{an ironseed.}

\item{quiet}{a logical indicating whether to silence messages.}
}
\value{
the previous of \code{.Random.seed} or \code{NULL}.
}
\description{
Initialize .Random.seed
}
\keyword{internal}
