% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ironseed-package.R
\docType{package}
\name{ironseed-package}
\alias{ironseed-package}
\title{Ironseed: Improved Random Number Generator Seeding}
\description{
A procedure for seeding R's built in random number generators using a variable-length sequence of values. Accumulates input entropy into a 256-bit hash digest or "ironseed" and is able to generate a variable-length sequence of output seeds from an ironseed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/reedacartwright/ironseed}
  \item Report bugs at \url{https://github.com/reedacartwright/ironseed/issues}
}

}
\author{
\strong{Maintainer}: Reed Cartwright \email{racartwright@gmail.com} (\href{https://orcid.org/0000-0002-0837-9380}{ORCID})

Other contributors:
\itemize{
  \item National Science Foundation DBI-1929850 [funder]
}

}
\keyword{internal}
